requirements.txt
# torch >= 0.4.1
# torchvision
numpy==1.19.5
argparse
scipy
opencv-python==3.4.2.17
opencv-contrib-python==3.4.2.17
matplotlib
imageio
tqdm
tensorflow==1.14.0
tensorboardX
tqdm
pyyaml
imageio
imgaug
jupyter
scikit-learn

torchgeometry
torchsummary
coloredlogs
